General progress.
This commit includes refinements to index.html's structure, as well as a number of advancements on the stylesheet.  It is beginning to vaguely resemble the target, but there are some roadblocks.  Namely, scaling is an issue.  The website shrinks too much to accomodate window size.  I need to figure out how to establish a baseline size for the elements.